Reference

This lists all functions included in the RAMS.jl module:

RAMS.RAMSDates — Method

RAMSDates(flist)

Takes array of file paths and returns datetimes.

Arguments

flist::Array{String,1}: 1D array of string file paths

Returns

dtarr::Array{DateTime,1}: 1D array of datetimes

RAMS.RAMSVar — Method

RAMSVar(flist, varname)

Function to read a variable from a list of RAMS data files.

Arguments

flist::Array{String,1}: 1D array of string file paths

varname::String: Name of variable

Returns

var::Array: Output variable

RAMS.dropmean — Method

dropmean(var, meandrop)

Function to drop dimensions used while taking a mean. Specify meandrop as a tuple of dimensions (e.g. meandrop=(1,2))

RAMS.list_files — Method

list_files(dir::String)

List all RAMS data files (*.h5) in dir. Effectively, this removes the *.txt header files from the built-in readdir() results.

RAMS.vert_int_4d — Method

vert_int_4d(var, ztn; notime=false)

Vertically integrate var along z axis with heights specified in ztn. Assumes that var has dimensions [t, x, y, z]

Specify notime=true to allow for vars with no t dimension (will also be omitted from int_var)

Returns:

int_var::Array: Integrated array with dimensions [t,x,y]