Reference

This lists all functions included in the RAMS.jl module:

RAMS.RAMSDates — Method

RAMSDates(flist)

Takes array of file paths and returns datetimes.

Arguments

flist::Array{String,1}: 1D array of string file paths

Returns

dtarr::Array{DateTime,1}: 1D array of datetimes

RAMS.RAMSVar — Method

RAMSVar(flist, varname)

Function to read a variable from a list of RAMS data files. If varname is a string, the specified variable will be returned. If varname is Vector{String}, the listed variables will be added together and returned.

Arguments

flist::Array{String,1}: 1D array of string file paths

varname::String: Name of variable.

Optional: varname::Vector{String}: Vector of variable names, the listed variables will be added together

Returns

var::Array: Output variable

RAMS.dropmean — Method

dropmean(var, meandrop)

Function to drop dimensions used while taking a mean. Specify meandrop as a tuple of dimensions (e.g. meandrop=(1,2))

RAMS.list_files — Method

list_files(dir::String)

List all RAMS data files (*.h5) in dir. Effectively, this removes the *.txt header files from the built-in readdir() results.

RAMS.mslp — Method

mslp(temp, press, height)

Calculate Mean Sea Level Pressure from temperature, pressure, and height

RAMS.pressure — Method

pressure(exner)

Return the pressure given the Exner function (PI in RAMS).

RAMS.rh — Method

rh(rv, theta, exner)

Calculates RH in percent (divide by 100 for fraction) Note: for this function THETA and PI must be arrays

RAMS.temperature — Method

temperature(theta, exner; celsius=false)

Returns temperature given potential temperature (THETA in RAMS) and the exner function * c_p (PI in RAMS). Specify celsius=true for celsius.

RAMS.vert_int — Method

vert_int(var, ztn; notime=false)

Vertically integrate var along z axis with heights specified in ztn using the Trapezoidal rule. Assumes that var has dimensions [x, y, z, t], [x, y, z], or [z, t].

Returns:

int_var::Array: Integrated array with dimensions [x,y,t]

RAMS.wsat — Method

wsat(theta, exner)

Saturation mixing ratio based on THETA and exner (PI).

Note: for this function THETA and PI must be arrays